Upgrade to latest shacl-tulip that fixes duplicate object values in formData
see commit 4d08a80492eb853b9247da2a89d3adc73fea87bb

This was picked up in shacl-vue's formEditor component, where duplicate values were
displayed for a single field. The duplicate values were inadvertently added when
calling quadsToFormData in the onBeforeMount hook of the formEditor component. That
call is necessary, but since it is called multiple times it had the unwanted result
of adding duplicates. This is also a general fix, since shacl-tulip should not be
adding duplicates to form data, whether intended or not.