Config option 'textTruncateWidth' and displaying literal strings
This introduces the new 'textTruncateWidth' config option which controls whether literal
strings are truncated or wrapped. To truncate it, the value can be set to a CSS width,
e.g. '50%' or '500px'. A number is cast as a string and interpreted as pixels. The default
is '85%'. To switch off this feature (text will be truncated by default), set the value to
false, which means all text will be wrapped at maximum width of the parent container.

To implement this, several changes were made to the LiteralNodeViewer, which had previously
supported a similar feature but by default (i.e. not configurable) and using a tooltip. The
LiteralNodeViewer will now, if text is truncated, display a button that allows users to
expand the full wrapped text, and also to truncate it again. The necessary props are passed
to this component from the NodeShapeViewer, which receives this configuration from the
existing 'componentConfig' composable.