Introduce a handlers and registerHandler per formEditor
This allows child components to inject 'registerHandler' and to then register a handler
function, of type 'save' or 'cancel'. These registered handlers will be called by an
editor component's parent 'formEditor', respectively when the save or cancel button is
pressed by the user. This is a uniform API that can be used by any editor component, and
the need for it initially came from the AssociationClassEditor. This prevents the need for
editors to run teardown/cleanup functions in their 'onBeforeUnmount', which might introduce
timing or reactivity issues in relation to other components in the hierarchy. With the new
functionality, all teardown code happens within the same component, but with contexts from
the originating components