Fix association class bugs related to teardown
A bug occurred during form save, where the association class record created by an 'AssociationClassEditor'
and that remained unused (i.e. no value selected/specified) was still linked to the parent record, and
hitting save and returning to the record viewer showed an empty association. This errored further when
the user attempts to hit the 'More details' button, because the association class record appeared not to
exist.

The cause was that, while the association class record was correctly removed on cancel, it was not done
during save. It was also noticed then, that a form save event did not account for the possibility (or
likelihood) of the association class field not being filled in. I.e. it is also necessary to check, on
save, if the field has a value and if not: both the association class record needs to be deleted from
the graph, and the pointer to that blank node ID from the current form's formData should also be removed.

Importantly, this needed to be done before the form is saved, because that action will take the latest
state of formData and turn it into quads and add them to the graph. To achieve this, this cleanup is
now done using the recently introduced 'registerHandler' from formData. Both 'save' and 'cancel' handlers
are registered. The formEditor component runs the save handler at the beginning of its save function,
before calling 'formData.saveNode'.

Another importan aspect here was that the removal of the association class record's blank node ID from
formData could not be done with shacl-tulip's removeObject function, since it uses splice to remove an
array element, given the index. And since this might be called for multiple indices of the same array,
indices of elements will change when an element is removed. So, instead, the relevant element is found
by '_key' and then the value is set to null. An element with a null value is ignored, intentionally, by
'formData.saveNode'.