Broaden class search for finding single item to list
This was needed because the search was using subject, predicate and object to find the relevant
quad in the graph. The subject is the node ID, the predicate is RDF.type, and the object was
the class IRI of the InstancesSelectEditor (props.node_uid). However, the selected item (from
modelValue) could also be a subclass of the class IRI. E.g. the class IRI could be Thing while
the RDF.type of the selected item could be Person. The initial restrictive search criteria would
then not find the specific quad, because the quad where that node ID has RDF.type Thing does
not exist. So we remove the object in order to still find the quad.