Introduce approach that does not lead to remounting 'FormEditor' on each view
There has been a long unadressed issue of the 'FormEditor's remounting behaviour, which causes
state to be reset upon remounting, even if the form was already opened before. This typically
happens when a user closes a different form that then brings the focus back to the current form,
which then leads to it remounting instead of just having maintained the previous state.

This commit fixes that issue with two concrete changes to the 'ShaclVue' component:
- use the eager prop on 'v-expansion-panel-text', which is the parent component of the 'FormEditor'
  and results in the 'FormEditor' instance remaining mounted
- wrap 'FormEditor' component in a 'v-show', which will only show the current form, while the others
  remain hidden yet still rendered; without this, the other open forms are also rendered which leads
  to bad UX

In response to this fix, some updated functionality for editor components was also necessary,
because code that used to run in 'onBeforeMount' and 'onMounted' hooks were now not running anymore
due to the 'FormEditor' not remounting. Specifically:
- The usual behaviour when canceling or saving a form was to run 'formData.removeSubject'. For editors
  that need to use said 'formData' upon returning to a form, specifically the 'AssociationClassEditor',
  this behaviour was fine, because it could recreate the 'formData' from graph quads on remounting. A
  different behaviour has now been introduced to conditionally run 'formData.removeSubject' based on
  an argument stored in 'openForms' when the form is initially opened. This argument, 'removeNode'
  can have one of the values 'true', 'false', 'onCancel', or 'onSave', and is passed to 'openForms' by
  the 'AssociationClassEditor' (via 'addForm' <- 'editInstanceItem') telling the 'formEditor' when
  exactly to remove the subject from 'formData'. The default of that argument is 'true'.
- Handlers were introduced in a07527a9414355c1aa3d4b64b6691fbc247661f9. Here we extend it to include
  'reopen' handler registration. These will run from inside a form if that form is reopened after it
  was collapsed for another form. To achieve this, the 'FormEditor' component has a new watcher to
  check for changes in the length of the 'openForms' ref, and if the length decreases (i.e. a form was
  closed) the currently open form will run any registered 'reopen' handlers. This functionally is used
  to recalculate related quads and show them in the information tooltip for an existing association
  class record, once the form that edits those relations is closed.

Lastly, a TODO: 88f402c51a1e0488d0d593d5b20b5cfe1b478109 introduced a workaround to the 'FormEditor'
remounting issue, specifically in order to have the 'show_all_fields' switch value can persist across
form opening/closing actions. This needs to be reverted.