Also fetch association class related records when rendering record
The 'NodeShapeViewer' used to get all related records of the current record from the graph, then loop
through them and if the quad object termtype was a named node, it would also fetch that specific record.
This is useful because such records are used in display name generation of related records. Now, since
we have started using association class (i.e. blank node) records more and more, the need was identified
for blank node records to also have display names generated. However, this could not be done because
records related to blank node records were not being fetched. This changes that, so that display names
can be generated one extra level of recursion deeper.