Improved blank node deduplication approach
The existing blank node deduplication approach removed blank nodes that were considered
redundant in the context of a root named node. However, this turned out to be too strict
because blank nodes of an identical structure but used in different hierarchical branches
stemming from the same root node were also deduplicated. This is not desired. We want to
deduplicate siblings, but not cousins.

This commit adds utility functions to allow a fingerprint calculation that is based on
both the existing structure hash, as well as a new context key. The context key for a
blank node record is determined by the path of the blank node (via predicate) from the
parent node. The 'parseTTLandDedup' function in the 'ReactiveRdfDataset' now use these
new utilities in order to decide which blank nodes to deduplicate.