Add ability to link associated record display name to keyInfoRole record
5faeda80dfbc12d70bc2e28a0897a658ffaee10e introduced support for displaying blank node records
via configured display name outside of the 'More Details' block. Until now, this was just text,
even though these displays referred to blank node records that linked actual PID records. More
specifically, this could be association class records with a 'keyInfoRole' configured, e.g. an
'attribution' record with the 'rdf:object' annotated as the 'keyInfoRole' property and pointing
to a 'Person' record PID.

This commit lets the 'NodeShapeViewer' collect the information related to those 'keyInfoRole'
properties, if they exist, and will render the text as a link, which will allow navigating to
those related records (the 'Person' in the example above) when clicked.